+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 - Modify some functions in `todoist` module to return structs instead of strings
 - Add `list deadline`
 - Add deadline to `task edit`
+- Sort tasks based on deadline
 
 
 ## 2025-04-30 v0.7.1